Moved ScanStatement
moved ScanStatement out of Interpreter.cs, functions.flow.cs and constants.cs into Btrieve.cs

removed "using WpfCSCS" from Interpreter.cs

Made ProcessBlock and SkipBlock functions(in Interpreter.cs) public
